requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
ibm-watsonx-ai
python-dotenv
